sequelizeCRUD

+++ models/sequelize.tasks.js
@@ -2,8 +2,10 @@ const Sequelize = require('sequelize');
 
 const sequelize = require('../db/sequelize.provider');
 
-// First model
 const TaskSchema = sequelize.define('task', {
+	userId: {
+		type: Sequelize.INTEGER
+	},
 	content: {
 		type: Sequelize.STRING
 	}
@@ -12,29 +14,21 @@ const TaskSchema = sequelize.define('task', {
 /*
 // in MySQL
 let exampleTask = {
-	id: 0, // default
-	content: 'Some text', // required
+	id: 1, // number integer, default
+	userId: 1, // number integer, handle required
+	content: 'Some text' // string, handle required
 };
 */
 
-// CRUD
+// task CRUD
 
 class Task {
 	constructor(model) {
 		this.model = model;
 	}
 
-	showList() {
-		return this.model.findAll()
-			.then(results => results)
-			.catch(err => {
-				console.error('Error read tasks from MySQL:\n', err);
-				return err;
-			});
-	}
-
-	addTask(content) {
-		return this.model.create({ content })
+	createTask(userId, content) {
+		return this.model.create({ userId, content })
 			.then(result => result)
 			.catch(err => {
 				console.error('Error write new task to MySQL:\n', err);
@@ -42,10 +36,17 @@ class Task {
 			});
 	}
 
-	changeTask(id, content) {
-		return this.model.update({ content }, {
-			where: { id }
-		})
+	readTasks(userId) {
+		return this.model.findAll({ where: { userId } })
+			.then(results => results)
+			.catch(err => {
+				console.error('Error read tasks from MySQL:\n', err);
+				return err;
+			});
+	}
+
+	updateTask(id, updates = {}) {
+		return this.model.update(updates, { where: { id } })
 			.then(result => result)
 			.catch(err => {
 				console.error('Error update task in MySQL:\n', err);
@@ -54,12 +55,19 @@ class Task {
 	}
 
 	deleteTask(id) {
-		return this.model.destroy({
-			where: { id }
-		})
-			.then(() => true)
+		return this.model.destroy({ where: { id } })
+			.then(result => true)
 			.catch(err => {
 				console.error('Error delete task from MySQL:\n', err);
+				return err;
+			});
+	}
+
+	deleteAllUserTasks(userId) {
+		return this.model.destroy({ where: { userId } })
+			.then(result => true)
+			.catch(err => {
+				console.error('Error delete user tasks from MySQL:\n', err);
 				return err;
 			});
 	}
